refactor: componentization for modal utilization
