+++ apps/web/src/common/modules/modals/multi-factor-authentication-modal/MultiFactorAuthenticationEmailModal.vue
@@ -5,20 +5,20 @@ import {
 import type { TranslateResult } from 'vue-i18n';
 
 import {
-    PButton,
     PButtonModal,
-    PCollapsibleToggle,
     PFieldGroup,
-    PI,
     PTextInput,
 } from '@spaceone/design-system';
 
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
-import { postEnableMfa, postValidationMfaCode } from '@/lib/helper/multi-factor-authentication-helper';
+import { postValidationMfaCode } from '@/lib/helper/multi-factor-authentication-helper';
 
 import { useProxyValue } from '@/common/composables/proxy-state';
+import CollapsibleContents
+    from '@/common/modules/modals/multi-factor-authentication-modal/modules/CollapsibleContents.vue';
+import EmailInfo from '@/common/modules/modals/multi-factor-authentication-modal/modules/EmailInfo.vue';
 
 interface Props {
     email?: string
@@ -64,16 +64,6 @@ const resetFormData = () => {
 };
 
 /* API */
-const handleClickSendEmailButton = async () => {
-    await postEnableMfa({
-        user_id: state.userId,
-        mfa_type: props.mfaType,
-        options: {
-            email: props.email,
-        },
-        domain_id: state.domainId,
-    });
-};
 const handleClickConfirmButton = async () => {
     state.loading = true;
     try {
@@ -108,23 +98,7 @@ const handleClickConfirmButton = async () => {
     >
         <template #body>
             <div class="modal-content-wrapper">
-                <div class="email-info-wrapper">
-                    <div class="sent-email-wrapper">
-                        <div class="contents-wrapper">
-                            <p>{{ $t('COMMON.MFA_MODAL.SENT_DESC') }}</p>
-                            <div class="email-wrapper">
-                                <p-i name="ic_envelope-filled"
-                                     height="0.875rem"
-                                     width="0.875rem"
-                                     color="inherit"
-                                />
-                                <p class="email-text">
-                                    {{ props.email }}
-                                </p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
+                <email-info :email="props.email" />
                 <p-field-group :label="$t('COMMON.MFA_MODAL.VERIFICATION_CODE')"
                                :invalid="validationState.isValidationCodeValid"
                                :invalid-text="validationState.validationCodeInvalidText"
@@ -136,23 +110,9 @@ const handleClickConfirmButton = async () => {
                                   @update:value="handleChangeInput"
                     />
                 </p-field-group>
-                <div class="collapsible-wrapper">
-                    <p-collapsible-toggle v-if="state.isCollapsed"
-                                          v-model="state.isCollapsed"
-                    >
-                        {{ $t('COMMON.MFA_MODAL.COLLAPSE_TITLE') }}
-                    </p-collapsible-toggle>
-                    <p v-if="!state.isCollapsed"
-                       class="collapsed-contents"
-                    >
-                        {{ $t('COMMON.MFA_MODAL.COLLAPSE_DESC') }}
-                        <p-button class="send-code-button"
-                                  @click.prevent="handleClickSendEmailButton"
-                        >
-                            <span class="emphasis">{{ $t('COMMON.MFA_MODAL.SEND_NEW_CODE') }}</span>
-                        </p-button>
-                    </p>
-                </div>
+                <collapsible-contents :mfa-type="props.mfaType"
+                                      :email="props.email"
+                />
             </div>
         </template>
         <template #confirm-button>
@@ -167,40 +127,8 @@ const handleClickConfirmButton = async () => {
         @apply flex flex-col;
         margin-bottom: 1rem;
 
-        .email-info-wrapper {
-            margin-bottom: 1rem;
-            .sent-email-wrapper {
-                @apply flex justify-between items-center bg-gray-100 rounded text-label-md text-gray-700;
-                padding: 0.5rem;
-                .contents-wrapper {
-                    @apply flex flex-col;
-                    .email-wrapper {
-                        @apply flex items-center font-bold;
-                        gap: 0.375rem;
-                    }
-                }
-            }
-        }
-
         .text-input {
             width: 100%;
-        }
-
-        .collapsible-wrapper {
-            margin-top: 2rem;
-            .collapsed-contents {
-                @apply text-paragraph-sm text-gray-500;
-                .send-code-button {
-                    @apply text-label-xs font-normal text-blue-700;
-                    height: 1rem;
-                    background: initial;
-                    padding: 0;
-                    margin: 0;
-                    .emphasis:hover {
-                        text-decoration: underline;
-                    }
-                }
-            }
         }
     }
 }
